project: renamed project dir
