+++ client/src/store/reducers/template.reducer.js
@@ -1,6 +1,10 @@
 import { TEMPLATE_GET_SUCCESS } from "../action.types";
 
-const onTemplateGetSuccess = (state, { templates }) => templates;
+import { normalizeLevel1 } from "../json.normalizr.js";
+
+const onTemplateGetSuccess = (state, { templates }) => {
+    return normalizeLevel1(templates, { entity: "templates" });
+};
 
 export default (state = null, { type, data }) => {
     switch (type) {
